Feature/endpoint orderproduct

+++ src/api/routes.py
@@ -90,7 +90,6 @@ def delete_categories():
 def populate_category():
     category_list = [u'Analgesico', u'Antibiotico', u'Dermatológico y cosmético', u'Nutrición', u'Pediátrico', 
                      u'Primeros auxilios', u'Salud digestiva', u'Tratamientos', u'Vitaminas']
-    #This list will be the first list just to populate the database, afterwards it will be redirected to the database to read the data
 
     for item in category_list:
         category = Category()
@@ -154,11 +153,7 @@ def populate_product():
                 image_url=med["image_url"],
                 description=med["description"]
             )
-            
-            
-
             db.session.add(product)
-
         db.session.commit()
         return jsonify({"message": "Products populated successfully"}), 200
 
@@ -166,26 +161,6 @@ def populate_product():
         print(e.args)
         db.session.rollback()
         return jsonify({"message": str(e)}), 500
-
-   
-    # for i in range(8):
-    #     num = i
-    #     product = Product()
-    #     product.generic_name = "Perifar" + ' ' + str(num)
-    #     product.active_ingredient = "Ibuprofeno"
-    #     product.category_id = category.category_id
-    #     product.price = 50
-    #     product.stock_quantity = 13
-    #     product.image_url = 'url'
-    #     product.description = 'A nice laugh the best medicine'
-    #     db.session.add(product)
-
-    # try: 
-    #     db.session.commit()
-    #     return jsonify("Adding product"), 200
-    # except Exception as error:
-    #     db.session.rollback()
-    #     return jsonify(f"{error}"), 500
 
 @api.route('/product', methods=['POST'])
 def add_product():
@@ -424,23 +399,6 @@ def get_one_order(theid = None):
             return jsonify({"message": "Order not Found"}), 404
     return jsonify({"message": "Id doesnt correspond to an order right now"}), 400
 
-@api.route('/order', methods=['POST'])
-@jwt_required
-def create_kart():
-    #Probablemente hay que reajustar esto cuando se cree la parte de logins y de jwt
-    user = User.query.get(get_jwt_identity())
-    data = request.json
-    if data is not None:
-        order = Order(user_id=user, order_status='Kart', order_type='Pickup')
-        try:
-            db.session.add(order)
-            return jsonify({"message": "Creating Kart"}), 201
-        except Exception as error:
-            print(error.args)
-            db.session.commit()
-            return jsonify({"message": "Problem commiting"}), 500
-    else:
-        return jsonify({"message": "Kart Couldnt be Created"}), 400
     
 @api.route('/order/<int:theid>', methods=['DELETE'])
 def delete_order(theid=None):
@@ -516,17 +474,18 @@ def get_reviews():
 def get_one_review(theid):
     review = Reviews()
     review = review.query.get(theid)
-
     if review is not None:
         return jsonify(review.serialize()), 200
     else:
         return jsonify({"message": "Review Id doesnt exist"}), 404
 
 @api.route('/review', methods=['POST'])
+@jwt_required()
 def create_review():
+    user = User.query.get(get_jwt_identity())
     data = request.json
     if data is not None:
-        review = Reviews(user_id=data['user_id'], product_id=data['product_id'], comment=data['comment'])
+        review = Reviews(user_id=user.user_id, product_id=data['product_id'], comment=data['comment'])
         db.session.add(review)
 
         db.session.commit()
@@ -535,14 +494,21 @@ def create_review():
         return jsonify({"message": "Review is not created"}), 500
     
 @api.route('/review/<int:theid>', methods=['DELETE'])
+@jwt_required()
 def delete_review(theid=None):
+    user=User.query.get(get_jwt_identity())
+
     if theid is not None:
-        review = Reviews()
-        review = review.query.get(theid)
-        if review is not None:
+        review = Reviews.query.get(theid)
+        if review is not None and user.user_id == review.user_id:
             db.session.delete(review)
-            db.session.commit()
-            return jsonify({"message": "Review Destroyed"}), 200
+            try:
+                db.session.commit()
+                return jsonify({"message": "Review Destroyed"}), 200
+            except Exception as error:
+                print(error.args)
+                db.session.rollback()
+                return jsonify({"message": f"{error.args}"}), 500
         else:
             return jsonify({"message": "Review Doesnt Exist"}), 404
     return jsonify({"message": "Id is None"}), 500        
@@ -560,7 +526,6 @@ def delete_reviews():
 
 @api.route('/review/populate', methods=['GET'])
 def populate_reviews():
-    #review_id, user_id, product_id, comment
     review = Reviews()
     
     user = User()
@@ -581,50 +546,24 @@ def populate_reviews():
         print(error.args)
         db.session.rollback()
         return jsonify(f"{error.args}"), 500
-
-
-@api.route('/order_product/<int:theid>', methods=['GET'])
-@jwt_required()
-def get_products_in_order(theid):
-    #Rehacer
-    order_product = OrderProduct()
-    order_product = order_product.query.all()
-    user_order_product = []
-
-    for op in order_product:
-        if op.order_id == theid:
-            user_order_product.append(op)
-      
-    return jsonify([item.serialize() for item in op]), 200
-
-@api.route('/order/<int:theid>', methods=['POST'])
-@jwt_required()
-def add_product_in_order():
-    # 1 - JWT required 
-    # 2 - Llega con request
-    # 3 - Verificar los endpoints de Order
     
-    # [Usuario Logeado] -> Order, Producto
-
-    user = User.query.get(get_jwt_identity())
-    if user is None:
-        return jsonify("User not found"), 404
-
-    data = request.json
-    
-    cart_order = Order.query.filter(Order.user_id == user, Order.order_status == "KART").one_or_none()
-
-    if cart_order is None:
-
-        #Crear una orden nueva aca
-
+def create_kart(user):
+    if user is not None:
+        order = Order(user_id=user.user_id, order_status='KART', order_type='PICKUP')
+        db.session.add(order)
+        try:
+            db.session.commit()
+            return order.order_id
+        except Exception as error:
+            print(error.args)
+            db.session.rollback()
+            return jsonify({"message": "Problem commiting"}), 500
     else:
-        pass
-        #agregar producto a la orden existente
-    pass
-
+        return jsonify({"message": "User doesnt Exist"}), 400
+    
 @api.route('/login', methods=['POST'])
 def login():
+
     data = request.json
     email = data.get("email", None)
     password = data.get("password", None)
@@ -632,23 +571,27 @@ def login():
     if email is None or password is None:
         return jsonify({"message": "Email and password required"}), 400
     else:
-        # Aca deberia chequear el pass, primero armar el acceso con salt
         user = User.query.filter_by(email=email).one_or_none()
 
         if user is None:
-            return jsonify({"message":"Incorrect Email"}), 400
+            return jsonify({"message":"Incorrect Email or Password"}), 400
         else:
             if check_password(user.password, password, user.salt):
                 token = create_access_token(identity=user.user_id)
                 return jsonify({"token":token}), 200
             else:
-                return jsonify({"message": "Bad Password"}), 400    
+                return jsonify({"message": "Bad Info"}), 400    
     
 @api.route('/update-password', methods=['PUT'])
 @jwt_required()
 def update_pass():
-    email = get_jwt_identity()
+    #Data que viene de entrada
+    user = get_jwt_identity()
     body = request.json
+
+    #Consiguiendo correo para actualizar la password
+    user = get_user(user)
+    email = user.get('email')
 
     user = User.query.filter_by(email=email).one_or_none()
 
@@ -665,3 +608,72 @@ def update_pass():
         except Exception as error:
             print(error.args)
             return jsonify("Password couldnt be updated"), 500
+
+@api.route('/order/<int:theid>', methods=['GET'])
+@jwt_required()
+def get_products_in_order(theid):
+    #Se pide que tenga jwt, para evitar que alguien revise las ordenes individuales de las personas ?
+    products_in_order = OrderProduct.query.filter_by(order_id=theid).all()
+    if products_in_order is None:
+        return jsonify({"Message": "There are no products in this order"}), 400
+    return jsonify([item.serialize() for item in products_in_order]), 200
+
+@api.route('/order/<int:theid>', methods=['POST'])
+@jwt_required()
+def add_product_in_order(theid):
+    user = User.query.get(get_jwt_identity())
+    if user is None:
+        return jsonify({"message": "User not found"}), 404
+
+    data = request.json
+    if not data or 'stock' not in data or not isinstance(data['stock'], int):
+        return jsonify({"message": "Invalid data: 'stock' is required and must be an integer"}), 400
+
+    order = Order.query.filter_by(user_id=user.user_id, order_status='KART').first()
+    
+    if order is None:
+        order_id = create_kart(user)
+    else:
+        order_id = order.order_id
+
+    order_product = OrderProduct(
+        order_id=order_id,
+        user_id=user.user_id,
+        product_id=theid,  
+        stock=data['stock']
+    )
+
+    db.session.add(order_product)
+
+    try:
+        db.session.commit()
+        return jsonify({"message": "Producto agregado a la orden"}), 201
+    except Exception as error:
+        db.session.rollback()
+        return jsonify({"message": f"Error: {error.args}"}), 500
+
+@api.route('/order/<int:theid>', methods=['DELETE'])
+@jwt_required()
+def delete_product_in_order(theid):
+    user = User.query.get(get_jwt_identity())
+    if user is None:
+        return jsonify("You need to log in"), 400
+    
+    order = Order.query.filter(Order.user_id==user.user_id, Order.order_status=='KART').one_or_none()
+    if order is not None:
+        product = Product.query.filter(OrderProduct.order_id==order.order_id, OrderProduct.product_id==theid)
+        if product is not None:
+            db.session.delete(product)
+        else:
+            return jsonify({"message": "This Product Does not Exist for this Order"}), 404
+    else:
+        return jsonify({"message": "Order Does Not Exist, something bad happened here"}), 400
+    
+    try:
+        db.session.commit()
+        return jsonify({"message": "Product has been deleted from this order"}), 200
+    except Exception as error:
+        print(error.args)
+        db.session.rollback()
+        return jsonify({"message": f"{error.args}"}), 500
+
